Day7 Progress
Topic of the the day:Java string introduction

In Java, string is basically an object that represents sequence of char values. An array of characters works same as Java string.

The elements of a String are called characters. The number of characters in a String is called the length, and it can be retrieved with the String.length() method. 

Question:

Given two strings of lowercase English letters, A and B, perform the following operations:



Sum the lengths of A and B.

Determine if is Alexicographically larger than B (i.e.: does A come before B in the dictionary?).

Capitalize the first letter in A and B and print them on a single line, separated by a space.



Input Format



The first line contains a string A. The second line contains another string B. The strings are comprised of only lowercase English letters.



Output Format



There are three lines of output:

For the first line, sum the lengths of A and B .

For the second line, write Yes if is A lexicographically greater than B otherwise print No instead.

For the third line, capitalize the first letter in both A and B and print them on a single line, separated by a space.

Logic:

1.Take input from the user

2.Compare the first string with another

3.Convert  lowercase to uppercase 

Input 1

abhinav agarwal

Output 

14

No

Abhinav

Agarwal

Input 2

hello

abhinav

Output 

12

Yes

Hello Abhinav