Day15 Progress
Topic of the Day:Java 1D array
An array is a simple data structure used to store a collection of data in a contiguous block of memory. Each element in the collection is accessed using an index, and the elements are easy to find because they're stored sequentially in memory.
Input Format
The first line contains a single integer,n , denoting the size of the array.
Each line i of the n subsequent lines contains a single integer denoting the value of element ai.

Output Format
You are not responsible for printing any output to stdout. Locked code in the editor loops through array  and prints each sequential element on a new line.

 Input
5
10
20
30
40
50

Output
10
20
30
40
50

Explanation
When we save each integer to its corresponding index in a, we get a=[10,20,30,40,50]. The locked code prints each array element on a new line from left to right.