Day23 Progress
Topic of the Day: Java Interface
A Java interface can only contain method signatures and fields. The interface can be used to achieve polymorphism. In this problem, you will practice your knowledge on interfaces.

You are given an interface AdvancedArithmetic which contains a method signature int divisor_sum(int n). You need to write a class called MyCalculator which implements the interface.

divisorSum function just takes an integer as input and return the sum of all its divisors. For example divisors of 6 are 1, 2, 3 and 6, so divisor_sum should return 12. The value of n will be at most 1000.

Input
6

Output
I implemented: AdvancedArithmetic
12

Explanation
Divisors of 6 are 1,2,3 and 6. 1+2+3+6=12.