Day26 Progress
Topic Of the Day: Java Syntax Checker
Using Regex, we can easily match or search for patterns in a text. Before searching for a pattern, we have to specify one using some well-defined syntax.

In this problem, we are given a pattern. we have to check whether the syntax of the given pattern is valid.

Note: In this problem, a regex is only valid if you can compile it using the Pattern.compile method.

Input Format
The first line of input contains an integer , denoting the number of test cases. The next  lines contain a string of any printable characters representing the pattern of a regex.

Output Format
For each test case, print Valid if the syntax of the given pattern is correct. Otherwise, print Invalid. Do not print the quotes.

Input
3
([A-Z])(.+)
[AZ[a-z](a-z)
batcatpat(nat

Output
Valid
Invalid
Invalid