Day34 Progress
Topic of the Day: Java Decode Ways (String)
Question: A message containing letters from A-Z can be encoded into numbers using the following mapping:

'A' -> "1"
'B' -> "2"
...
'Z' -> "26"
To decode an encoded message, all the digits must be grouped then mapped back into letters using the reverse of the mapping above (there may be multiple ways). For example, "11106" can be mapped into:
"AAJF" with the grouping (1 1 10 6)
"KJF" with the grouping (11 10 6)
Note that the grouping (1 11 06) is invalid because "06" cannot be mapped into 'F' since "6" is different from "06". Given a string s containing only digits, return the number of ways to decode it. The test cases are generated so that the answer fits in a 32-bit integer.

Example 1:
Input: s = "12"
Output: 2
Explanation: "12" could be decoded as "AB" (1 2) or "L" (12).

Example 2:
Input: s = "226"
Output: 3
Explanation: "226" could be decoded as "BZ" (2 26), "VF" (22 6), or "BBF" (2 2 6).

Testcase:
Your input:
"12"
"226"
"06"
Output:
2
3
0
Expected:
2
3
0

Explanation:
This commit does not belong to any branch on this repository, and may belong to a fork outside of the repository. Given a non-empty string containing only digits, determine the total number of ways to decode it. Explanation: It could be decoded as "AB" (1 2) or "L" (12).