Day72 Progress
Problem: An integer array nums sorted in non-decreasing order, return an array of the squares of each number sorted in non-decreasing order.
 
Example 1:
Input: nums = [-4,-1,0,3,10]
Output: [0,1,9,16,100]
Explanation: After squaring, the array becomes [16,1,0,9,100].
After sorting, it becomes [0,1,9,16,100].

Example 2:
Input: nums = [-7,-3,2,3,11]
Output: [4,9,9,49,121]

Testcase:
Your input
[-4,-1,0,3,10]
[-7,-3,2,3,11]
Output
[0,1,9,16,100]
[4,9,9,49,121]
Expected
[0,1,9,16,100]
[4,9,9,49,121]

Explanation:
Numbers (of objects) that can be arranged into a square array are called “square numbers. Square arrays must be full if we are to count the number as a square number. Here, 12 pennies are arranged in a square, but not a full square array, so 12 is not a square number. The number 12 is not a square number.