Day75 Progress
Problem: The root of a binary tree, return the postorder traversal of its nodes' values.

Example 1:
Input: root = [1,null,2,3]
Output: [3,2,1]

Example 2:
Input: root = []
Output: []

Testcase:
Accepted
Input
root =
[1]
Output
[1]
Expected
[1]

Explanation:
Postorder traversal is used to visit the node in the tree. It pursues the rule of LRN, which means Left-right-node. Basically, this traversal helps in getting the postfix phrase of a tree.